+++ src/os/mod.rs
@@ -12,4 +12,5 @@
 //! consistent abstraction.
 
 pub mod battery;
+pub mod cpu_freq;
 pub mod cpu_temp;