+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- A block "cpu_freq" for OpenBSD, that displays the frequency of the CPU in MHz
+  or GHz. If several CPUs are combined in one block, it's possible to display the
+  average or maximum frequency.
 - A block "cpu_temp" for OpenBSD, that displays the temperature of the CPU in
   Celsius or Fahrenheit.
 - Documentation in the [./docs](./docs) directory to explain the choices that